rename all spec file by replacing . by -
