+++ src/app/common/auth/action/auth-action.spec.js
