+++ src/app/common/auth/reducer/auth-reducer.spec.js
