+++ src/app/common/auth/service/auth-service.spec.js
