+++ src/app/common/component/avatar/user-avatar-component.spec.js
