+++ src/app/common/component/resource/resource-icon-component.js
