+++ src/app/common/component/resource/resource-icon-component.spec.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {shallow} from 'enzyme'
 import {expect} from 'chai'
-import ResourceIcon from './resource-icon.component'
+import ResourceIcon from './resource-icon-component'
 import LucioleSVG from './../../core/component/luciole-svg-component'
 
 describe('ResourceIcon', () => {
