+++ src/app/common/component/wip/wip-component.js
