+++ src/app/common/core/component/luciole-svg-component.spec.js
