+++ src/app/common/helper/mock-helper.spec.js
