+++ src/app/module/about/container/about-container.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import WorkInProgress from './../../../common/component/wip/wip.component'
+import WorkInProgress from './../../../common/component/wip/wip-component'
 
 /**
  * AboutPage container, used to define the composition of the AboutPage screen
