+++ src/app/module/buildings/container/buildings-container.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import WorkInProgress from './../../../common/component/wip/wip.component'
+import WorkInProgress from './../../../common/component/wip/wip-component'
 
 /**
  * BuildingsPage container, used to define the composition of the BuildingsPage screen
