+++ src/app/module/buildings/container/buildings-container.spec.js
@@ -1,24 +1,24 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import { expect } from 'chai'
-import { ContactPage } from './contact-container'
-import WorkInProgress from './../../../common/component/wip/wip.component'
+import { BuildingsPage } from './buildings-container'
+import WorkInProgress from './../../../common/component/wip/wip-component'
 
-describe('ContactPage', () => {
+describe('BuildingsPage', () => {
   describe('mapDispatchToProps', () => {
     it('Expect to return empty object', () => {
-      expect(ContactPage.mapDispatchToProps()).to.be.empty
+      expect(BuildingsPage.mapDispatchToProps()).to.be.empty
     })
   })
 
   describe('mapStateToProps', () => {
     it('Expect to return empty object', () => {
-      expect(ContactPage.mapStateToProps()).to.be.empty
+      expect(BuildingsPage.mapStateToProps()).to.be.empty
     })
   })
 
   describe('render', () => {
-    const wrapper = shallow(<ContactPage />)
+    const wrapper = shallow(<BuildingsPage />)
     it('Expect to contain a WorkInProgress ', () => {
       expect(wrapper.find(WorkInProgress)).to.be.length(1)
     })
