+++ src/app/module/contact/container/contact-container.spec.js
@@ -1,24 +1,24 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import { expect } from 'chai'
-import { BuildingsPage } from './buildings-container'
-import WorkInProgress from './../../../common/component/wip/wip.component'
+import { ContactPage } from './contact-container'
+import WorkInProgress from './../../../common/component/wip/wip-component'
 
-describe('BuildingsPage', () => {
+describe('ContactPage', () => {
   describe('mapDispatchToProps', () => {
     it('Expect to return empty object', () => {
-      expect(BuildingsPage.mapDispatchToProps()).to.be.empty
+      expect(ContactPage.mapDispatchToProps()).to.be.empty
     })
   })
 
   describe('mapStateToProps', () => {
     it('Expect to return empty object', () => {
-      expect(BuildingsPage.mapStateToProps()).to.be.empty
+      expect(ContactPage.mapStateToProps()).to.be.empty
     })
   })
 
   describe('render', () => {
-    const wrapper = shallow(<BuildingsPage />)
+    const wrapper = shallow(<ContactPage />)
     it('Expect to contain a WorkInProgress ', () => {
       expect(wrapper.find(WorkInProgress)).to.be.length(1)
     })
