+++ src/app/module/homepage/container/homepage-container.spec.js
@@ -4,7 +4,7 @@ import { expect } from 'chai'
 import { HomePage } from './homepage-container'
 import HomePageLoggedIn from './../component/homepage-logged-in-component'
 import HomePageLoggedOff from './../component/homepage-logged-off-component'
-import WorkInProgress from './../../../common/component/wip/wip.component'
+import WorkInProgress from './../../../common/component/wip/wip-component'
 
 describe('HomePage', () => {
   describe('mapDispatchToProps', () => {
