+++ src/app/module/main/component/main-sidebar-burger-component.spec.js
