+++ src/app/module/main/component/main-user-card-component.spec.js
