+++ src/app/module/main/reducer/main-reducer.spec.js
