+++ src/app/module/main/service/user-service.spec.js
