+++ src/app/module/sidebar/action/sidebar-action.spec.js
