+++ src/app/module/sidebar/component/sidebar-block-resources-component.spec.js
