+++ src/app/module/sidebar/component/sidebar-line-resources-component.js
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react'
-import ResourceIcon from './../../../common/component/resource/resource-icon.component'
+import ResourceIcon from './../../../common/component/resource/resource-icon-component'
 import LucioleComponent from './../../../common/core/abstract/luciole-component'
 import ReactTooltip from 'react-tooltip'
