+++ src/app/module/sidebar/component/sidebar-link-component.spec.js
