+++ src/app/module/sidebar/component/sidebar-logged-in-component.spec.js
