+++ src/app/module/sidebar/component/sidebar-logged-off-component.spec.js
