+++ src/app/module/sidebar/component/sidebar-logo-component.spec.js
