+++ src/app/module/sidebar/container/sidebar-container.spec.js
