+++ src/app/module/sidebar/reducer/sidebar-reducer.spec.js
