+++ src/app/module/sidebar/service/resource-service.spec.js
